Env var key is static & add disable instructions
To ensure we're not misspelling the key name somewhere in the output I made it a static. I also added notes on how to disable the feature consistent with how it currently works (and also how all env var based enable/disable flags work in the existing buildpack I.e. they check for presence, not for content. This is also consistent with the Ruby language as `0`, `"0"`, and `"false"` (string false) are all "truthy". The only "falsey" values in Ruby are `false` (boolean false
) and `nil`