Address process booting determinism
The method we're using to launch a daemon process does not give us guarantees that `agentmon_loop --path` will be executed by the time the process boots. To address this we can loop and re-check for some period of time.

This method utilizes the output log from agentmon loop as directed by the spawn daemon command.