Report io error when detect cannot read a file. (#243)
It's best practice to use `try_exists()` instead of `exists()` when checking if a file exists or not because if there's a problem with the file (such as a permissions issue) then the check might fail even though the file is on disk. This seems like not a big deal, but can be very confusing when telling users "we didn't do X because you are missing file Y" but when they debug they see file Y and don't realize that the real problem isn't "missing file" but it's the hidden error preventing `exists()` from returning a true.


This commit uses fs_err_try_exists() from fs-err library that was introduced in: https://github.com/andrewhickman/fs-err/pull/48.

There are a few other places where `exists` is still called, those are nested in areas that don't already return a Result so they're harder to swap.