Use `pack build --force-color` in CI instead of `script` TTY workaround
As of Pack 0.33.0 (which this repo is using thanks to #462), there is a
new `--force-color` argument that makes `pack build` output logs
in colour even if a TTY wasn't detected (as is the case on GHA):
https://github.com/buildpacks/pack/releases/tag/v0.33.0

This makes the `script -e -c` TTY workaround in CI unnecessary.

I've also:
 - removed a redundant `cargo build` step (run does a build, and
   the Cargo compilation output can be suppressed using `--quiet`)
- replaced a manual git clone with `actions/checkout`, since it
  handles things like retries, faster shallow clones etc.

GUS-W-14968745.