Switch to the `zlib` flate2 backend
The default `miniz_oxide` flate2 backend has poor performance in debug/under QEMU:
https://github.com/rust-lang/flate2-rs/issues/297

Ideally we'd use the fastest `zlib-ng` backend, however it fails to cross-compile:
https://github.com/rust-lang/libz-sys/issues/93

As such we have to use the next best alternate backend, which is `zlib`.

This makes the `flate2` usage in this repo consistent with `libherokubuildpack`,
Python CNB, PHP CNB etc.