Switch to the new GitHub Actions runner groups
To update to Ubuntu 24.04, and to prevent the rate limit issues
mentioned here:
https://github.com/heroku/buildpacks-go/pull/283

I've not added the run image pre-pulling step used by some of
the other CNB repos, since this repo doesn't currently use a matrix
to test builder image versions, and instead varies the builder test
by test, which (a) means I'd have to hardcode a list, (b) means that
the benefit is less.

See also:
https://salesforce.quip.com/bu6UA0KImOxJ#temp:C:GZR57becb9df8d94f80b132168fd
https://github.com/actions/runner-images/blob/main/images/ubuntu/Ubuntu2404-Readme.md

GUS-W-16238120.