Use Vec instead of Option<Vec>
Discussion https://github.com/heroku/buildpacks-ruby/pull/325#discussion_r1772388102. Some([]) (empty vec) is a possible state from the type signature `Option<Vec<String>>` the state of Some/None is ambiguous without more information from the function. This code is easier to accidentally use the wrong logic in the consumer, but the correct implementation checks the direct properties we desire instead of relying on internal implementation of a function.

We could alternatively introduce a non-empty vector, there are several crates that provide this, but it's more overhead that needed here.