Move layer caching logic into shared location
The logic of Ruby install layer is generic enough based on the addition of a new trait bounds MetadataDiff that we can standardize on that trait and `TryMigrate`. This logic can then be re-used in all other layers to guarantee formatting consistency and cache behavior. We can later add unit tests for metadata differences to ensure they would cause a cache clear without needing to write full unit tests.