Use builder pattern for cache state
The `with_data` input is required due to type inference compilation requirements. If someone doesn't care they can call it with `|_,_|, ()` and it's explicit that they're not relying on the prior cache state. The alternative would be returning the old metadata always and then allowing the receiver to do something with it. You can still do that via `|old, _| old.clone()` as seen here, but it doesn't add a `Clone` requirement to the metadata.