Test file copying preserves mtime
Adds a test to ensure that a file copied via the AppCache feature preserves the file's original mtime. 


In #302 I determined that as long as the files copied into the cache dir preserve the original mtime then the FIFO cache cleaning logic that we rely on is valid. This test closes out this comment that asserted we need to verify whether or not the buildpack preserved mtime when copying files https://github.com/heroku/buildpacks-ruby/issues/302#issuecomment-2415128173.


[#close 302]