Add cp_r to dependencies
It appears the behavior of fs-extra with regard to mtime for copying/moving directories is system dependent. To compensate, I switched to `cp_r` which explicitly is designed to preserve mtime https://crates.io/crates/cp_r. It has ~27,000 lifetime downloads and has activity as recently as a 9 days ago. https://docs.rs/cp_r/latest/cp_r/#release-history

Copying and preserving mtime is the primary reason this library exists. From the docs.rs:

> Copy a directory tree, including mtimes and permissions.

It does not have the ability to skip-overwriting existing files which I need, however it has a generic `filter` closure which allows me to assemble this behavior myself.