Fix `docker run` with arguments (#404)
Fixes an issue where `docker run` with a command was not working correctly. Before:

```
$ docker run --rm --entrypoint /cnb/lifecycle/launcher sample-app "echo 'hi'"
hi
⛄️ 3.4.2 🚀 /Users/rschneeman/Documents/projects/work/buildpacks-ruby (main)
$ docker run --rm sample-app "echo 'hi'"
bash: echo 'hi': No such file or directory
$ pack inspect sample-app | grep Processes: -A10
Processes:
  TYPE                 SHELL        COMMAND        ARGS                                                                                         WORK DIR
  web (default)                     bash           -c bundle exec rackup --host "[::]" --port "${PORT:?Error: PORT env var is not set!}"        /workspace
```

After:

```
$ cargo libcnb package && pack build sample-app \
  --buildpack packaged/x86_64-unknown-linux-musl/debug/heroku_ruby \
  --path buildpacks/ruby/tests/fixtures/default_ruby --clear-cache
$ docker run --rm sample-app "echo 'hi'"
hi
```

The root cause is that when you `docker run` something, it uses the default command and uses the arguments you provide as arguments to that command. In this case we want the full shell, so the command needs to be `bash -c <arguments>` instead of `bash <arguments>`.