Remove error line wrapping
It looks awkward when the line is longer than the terminal and there is a leading `remote:` from a `git push` as the text does not stay in front of the `remote: ` it ends up looking something like this:

```
remote: ! Error: `Gemfile found with error
remote: !
remote: ! There was an error trying to read the contents of the application's Gemfile. The buildpack 
cannot continue if the Gemfile is unreadable.
remote: !
remote: ! Error info
```

We cannot prevent this from happening always, but we also don't have to purposefully have very long lines in our output.