Switch to the bullet_stream function interface (instead of type-state "streaming" interface) (#419)
* Use bullet_stream function interface

* Use bullet_stream function interface

* Use bullet_stream function interface

* Change: Use full command in the output

There's no real reason to hide the extra flags.

* Use bullet_stream function interface

* Use bullet_stream function interface

* Use bullet_stream function interface

* Use bullet_stream function interface

* Use bullet_stream function interface

* Use bullet_stream function interface

* Use bullet_stream function interface

* Remove bullet argument from interface

* Remove bullet argument from interface

* Remove bullet argument from interface

* Remove bullet argument from interface

* Remove bullet argument from interface

* Fix clippy warnings

* Remove bullet argument from interface

* Remove bullet argument from interface

* Remove bullet argument from interface

* Remove bullet argument from interface

* Remove bullet_stream::Print import

* Fix command representation

* Update example

* Use bullet_stream function style

* Remove error line wrapping

It looks awkward when the line is longer than the terminal and there is a leading `remote:` from a `git push` as the text does not stay in front of the `remote: ` it ends up looking something like this:

```
remote: ! Error: `Gemfile found with error
remote: !
remote: ! There was an error trying to read the contents of the application's Gemfile. The buildpack 
cannot continue if the Gemfile is unreadable.
remote: !
remote: ! Error info
```

We cannot prevent this from happening always, but we also don't have to purposefully have very long lines in our output.

* Changelog

* Fix clippy

* Prefer `call` over `handle` for interface name

This matches Ruby convention for Procs, lambdas, and service objects. Basically anything that has a generic execute this code" entrypoint usually uses `call`.