Replace deprecated struct with toml::Value
This took forever to debug, but basically serde doesn't play nice with deprecated structs https://github.com/serde-rs/serde/issues/2195 and unfortunately the warning isn't clear that it comes from generated code.

That leaves two options: Either use a top level `![allow(deprecated)]` in the file or remove the deprecated struct.

## Option: Top level `![allow(deprecated)]`

This option means we might accidentally miss some other deprecation notice in this (large) file. To avoid that we could refactor all of the metadata into their own module.

Alternatively

## Option: Replace the deprecated struct

In this case, we're no longer using the logic associated with this struct, we only need to keep it around for serialization/deserialization to support old cache data. When the data is migrated to the latest (v4) the field is ignored anyway, so all it has to do is not fail on serialization/deserialization. To that end, we can use `toml::Value` which holds any valid toml value.

## Decision: Replace with toml::Value

This option allows us to fully get rid of this struct so we can remove it independently of how long MetadaV1 (etc.) are around. The only issue is that it does not implement `Eq` but apparently that trait wasn't being used in our code any longer. (Previously, `Eq` was being used for cache invalidation.) 

The one test here that's updated validates that the cache is cleared when moving from `force_bundle_install_key=v1` to any other value. This is because the cache with that structure holds an older disk structure we're no longer using and we need to clear the contents to ensure extra files don't persist indefinetly.