Remove MetadataV1 and V2 from bundle install
Metadata structs are preserved so they can be "migrated" using the `TryMigrate` trait definition. This allows us to make changes to the metadata struct without accidentally busting the cache (which forces a re-install of all customer gems), which is slow and undesirable. They are introduced in sequence so V1 was the first ever struct etc.

This historical chain of structs is needed because we don't know when the last time someone deployed with this code i.e. we cannot guarantee they built their code with the last cache structure, it could have been months or years ago (in theory).

This change removes v1 and v2 structs as they were replaced before Fir was GA in April 14th, 2025 https://devcenter.heroku.com/changelog-items/3188.

V3 was introduced in https://github.com/heroku/buildpacks-ruby/pull/370 on December 13th, 2025. This means that any code built on Fir uses this cache structure, so we can safely remove the structs before v3.