Run `bundle install` on every push (#425)
* Always run `bundle install`

Previously: We inspected the state of the last `Gemfile.lock` and user environment variables to only run `bundle install` when something changed. This provides minimal value in practice and requires a lot of plumbing.

This change: Calls `bundle install` on every run, but does not clean up all code.

* Remove generating InstallState from cache logic

* Remove unused code

* Remove unused imports

* Introduce MetadataV4

* Pin function args to non-aliased type

* Remove warning as bundle install is always run

* Update type alias

Moving from v3 to v4 drops two fields.

* Remove unused code

* Remove error variant

* Remove unused code

* Deprecate MetadataDigest

* Replace deprecated struct with toml::Value

This took forever to debug, but basically serde doesn't play nice with deprecated structs https://github.com/serde-rs/serde/issues/2195 and unfortunately the warning isn't clear that it comes from generated code.

That leaves two options: Either use a top level `![allow(deprecated)]` in the file or remove the deprecated struct.

## Option: Top level `![allow(deprecated)]`

This option means we might accidentally miss some other deprecation notice in this (large) file. To avoid that we could refactor all of the metadata into their own module.

Alternatively

## Option: Replace the deprecated struct

In this case, we're no longer using the logic associated with this struct, we only need to keep it around for serialization/deserialization to support old cache data. When the data is migrated to the latest (v4) the field is ignored anyway, so all it has to do is not fail on serialization/deserialization. To that end, we can use `toml::Value` which holds any valid toml value.

## Decision: Replace with toml::Value

This option allows us to fully get rid of this struct so we can remove it independently of how long MetadaV1 (etc.) are around. The only issue is that it does not implement `Eq` but apparently that trait wasn't being used in our code any longer. (Previously, `Eq` was being used for cache invalidation.) 

The one test here that's updated validates that the cache is cleared when moving from `force_bundle_install_key=v1` to any other value. This is because the cache with that structure holds an older disk structure we're no longer using and we need to clear the contents to ensure extra files don't persist indefinetly.

* Update comment

* Document when struct was introduced

* Remove MetadataV1 and V2 from bundle install

Metadata structs are preserved so they can be "migrated" using the `TryMigrate` trait definition. This allows us to make changes to the metadata struct without accidentally busting the cache (which forces a re-install of all customer gems), which is slow and undesirable. They are introduced in sequence so V1 was the first ever struct etc.

This historical chain of structs is needed because we don't know when the last time someone deployed with this code i.e. we cannot guarantee they built their code with the last cache structure, it could have been months or years ago (in theory).

This change removes v1 and v2 structs as they were replaced before Fir was GA in April 14th, 2025 https://devcenter.heroku.com/changelog-items/3188.

V3 was introduced in https://github.com/heroku/buildpacks-ruby/pull/370 on December 13th, 2025. This means that any code built on Fir uses this cache structure, so we can safely remove the structs before v3.

* Remove Metadat v1 and v2 from ruby install

V3 was introduced December 13, 2025 before Fir went GA.

* Fix imports

* Remove unused code

* Update docs

* Remove TargetId::from_stack

This function was used to migrate metadata when we previously stored "heroku-24" instead of `ubuntu 24.04` in the cache. Now that cache structure is no longer supported, we can remove this logic

* Clippy

* Update tests

* Changelog

* Update custom cache diff function

Rename the cache diff function to align with its usage. Add comments explaining why it was added.

* Move Metadata type alias to the top