Configure Node CNB to skip pruning (#439)
* Configure Node CNB to skip pruning

The Node.js CNB is often used by the Ruby CNB to handle asset compilation for Rails apps. A recent change to the Node.js CNB made it so that dev dependencies are pruned at the end of its build phase, but these dev dependencies are then needed when the Ruby CNB wants to do asset compilation.

To support this use case, the Node.js CNB exposes a buildplan configuration that can control how some of these features work. This PR includes that required buildplan whenever Node.js tooling is required + modifies an integration test to verify that the configuration is working.

* Update CHANGELOG.md

Signed-off-by: Colin Casey <casey.colin@gmail.com>

* Only include Node.js buildplan config if `yarn` is required.

---------

Signed-off-by: Colin Casey <casey.colin@gmail.com>